Hw-2/Homework 2/2.1(stack)/2.1(Stack).Test/StackTest.cs
using System;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using _2._1_stack_;

namespace _2._1_Stack_.Test
{
    [TestClass]
    public class StackTest
    {
        [TestInitialize]
        public void Initialize()
        {
            stack = new Stack();
        }

        [TestMethod]
        public void PushTest()
        {
            stack.Push(1);
            Assert.IsFalse(stack.IsEmpty());
        }

        [TestMethod]
        public void PopTest1()
        {
            stack.Push(1);
            Assert.AreEqual(1, stack.Pop());
        }

        [TestMethod]
        public void PopTest2()
        {
            stack.Push(1);
            stack.Push(2);
            Assert.AreEqual(2, stack.Pop());
            Assert.AreEqual(1, stack.Pop());
        }

        [TestMethod]
        [ExpectedException(typeof(DeleteException))]
        public void PopTest3()
        {
            stack.Pop();
        }

        private Stack stack;
    }
}
